+++ Web/App/controllers/account/externalLoginController.js
@@ -38,7 +38,7 @@
                     userFactory.getAccessToken('', '', tempToken)
                         .success(function (tokenData) {
 
-                            // Actually this is not necessary, since search('tempToken', null) will redirect and 
+                            // Actually this is not necessary, since search('tempToken', null) will redirect and ...
                             vm.isAuthenticated = true;
 
                             // Clear search param